pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "handful"
version = "0.1.0"
description = "Real-time hand tracking and visualization"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy",
    "opencv-python",
    "mediapipe",
    "flask",
    "click",
    "pyyaml",
]


[project.scripts]
handful = "main:main"
handful-cli = "handful.cli.main:cli"

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "black",
    "isort",
    "mypy",
]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.10"
strict = true
